Retire multi-LoRA v1 and keep the slot mechanics
Removes the dataset-driven v1 backend end to end: the fully-async driver,
adapter controller and HTTP API, per-adapter data sources, the rollout data
plane (Sample.adapter routing, per-adapter metrics scoping), the update-path
push selection (loaded_adapters / pending-push bookkeeping), the v1 scheduler
and periodic adapter checkpoints, the controller CLI flags, and the v1 docs,
example, and tests.

Keeps the slot mechanics under megatron_utils/lora/: per-slot LayerWise Adam
(with the current LayerWise base-torch-optimizer contract and grad-norm API),
slot lifecycle, shard naming and rank slicing, plus MoE expert adapter support.
The weight-update path keeps its LoRA stream; adapters no longer ride the base
sync and the exported adapter handle flattens to AdapterSpec(slot, rank, alpha).
