+++ miles/backends/megatron_utils/checkpoint.py
@@ -15,7 +15,7 @@ from megatron.training.global_vars import get_args
 from miles.utils import megatron_bridge_utils
 from miles_plugins.models.deepseek_v4.arguments import assert_checkpoint_is_current, is_dsv4_model
 
-from .lora_utils import is_lora_enabled, is_lora_model, load_lora_adapter, save_lora_checkpoint
+from .lora.utils import is_lora_enabled, is_lora_model, load_lora_adapter, save_lora_checkpoint
 from .model_provider import LinearForLastLayer
 
 try:
