+++ miles/backends/training_utils/log_utils.py
@@ -210,9 +210,6 @@ def log_rollout_data(rollout_id: int, args: Namespace, rollout_data: RolloutBatc
                 "num_rollouts",
                 "n_adapters",
                 "adapter_slots",
-                "step_slots",
-                "step_adapter_names",
-                "step_adapter_batch_sizes",
                 "prompt_group_sizes",
             ]:
                 continue
