+++ tests/fast/ray/rollout/test_metrics.py
@@ -11,7 +11,7 @@ from miles.ray.rollout.metrics import (
     _compute_zero_std_metrics,
     log_rollout_data,
 )
-from miles.utils.types import AdapterRef, Sample, WeightVersionSpan, WeightVersionsPerCall
+from miles.utils.types import Sample, WeightVersionSpan, WeightVersionsPerCall
 
 
 class TestEpisodeResponseLengthMetrics:
@@ -59,18 +59,6 @@ class TestEpisodeResponseLengthMetrics:
 
         assert out["episode_response_length/mean"] == pytest.approx(4.5)
         assert out["episode_total_response_length/mean"] == pytest.approx(8.0)
-
-    def test_multi_lora_samples_emit_no_episode_length_metrics(self):
-        samples = [
-            make_sample(index=0, rollout_id=10, adapter=AdapterRef(name="adapter-a", slot=0)),
-            make_sample(index=0, rollout_id=10, adapter=AdapterRef(name="adapter-b", slot=1)),
-        ]
-
-        assert _compute_episode_response_length_metrics(samples) == {}
-        out = _compute_metrics_from_samples(make_args(advantage_estimator="ppo"), samples)
-        assert not any(key.startswith("episode_response_length/") for key in out)
-        assert "episode_total_response_length/mean" not in out
-        assert out["response_len/mean"] == pytest.approx(4.0)
 
     def test_removed_sample_has_zero_effective_length_but_keeps_total_length(self):
         sample = make_sample(
@@ -120,21 +108,6 @@ class TestTrainingSampleMetrics:
             make_sample(group_index=0, index=0, rollout_id=10, reward=1.0),
             make_sample(group_index=0, index=0, rollout_id=10, reward=1.0),
             make_sample(group_index=1, index=1, rollout_id=10, reward=0.0),
-        ]
-
-        out = _compute_training_sample_metrics(args, samples)
-
-        assert out["episode_raw_reward"] == pytest.approx(0.5)
-
-    def test_rollout_ids_are_scoped_by_adapter(self):
-        args = make_args(reward_key=None)
-        adapter_a = AdapterRef(name="adapter-a", slot=0)
-        adapter_b = AdapterRef(name="adapter-b", slot=1)
-        samples = [
-            make_sample(group_index=0, rollout_id=10, adapter=adapter_a, reward=1.0),
-            make_sample(group_index=0, rollout_id=10, adapter=adapter_a, reward=1.0),
-            make_sample(group_index=0, rollout_id=10, adapter=adapter_a, reward=1.0),
-            make_sample(group_index=0, rollout_id=10, adapter=adapter_b, reward=0.0),
         ]
 
         out = _compute_training_sample_metrics(args, samples)
